+++ extensions/vscode-aml/.gitignore
@@ -1,3 +1,4 @@
 dist
 node_modules
+local
 *.vsix
